Renamed some functions and arguments to more logical names that describe their purposes better. clean_text is now produced from a closure that maps a function over a dictionary in a specific way. It actually modifies the copy of the dictionary that is passed in currently, but perhaps it should leave the original untouched and return a modified duplicate. The output file now has headers written as the first line.
dict_signaturizer(signature_keys) is a closure that returns a function that will create a signature for any dictionary. A signature is defined as a tuple of the values contained in the dict at each key in signature_keys. For my purposes, a tuple of the cleaned 'address1' and 'zip' values will be sufficient to filter out most any duplicate entry.

duplicate_filter(collection, signature) has had its values renamed to better signify their purpose.